ZoneDraft to interface. #942
