+++ sphere-models/src/main/java/io/sphere/sdk/zones/ZoneDraft.java
@@ -1,5 +1,6 @@
 package io.sphere.sdk.zones;
 
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.neovisionaries.i18n.CountryCode;
 
 import javax.annotation.Nullable;
@@ -7,41 +8,27 @@ import java.util.Set;
 
 import static java.util.stream.Collectors.toSet;
 
-public class ZoneDraft {
-    private final String name;
+@JsonDeserialize(as = ZoneDraftImpl.class)
+public interface ZoneDraft {
     @Nullable
-    private final String description;
-    private final Set<Location> locations;
+    String getDescription();
 
-    private ZoneDraft(final String name, final Set<Location> locations, @Nullable final String description) {
-        this.description = description;
-        this.name = name;
-        this.locations = locations;
+    Set<Location> getLocations();
+
+    String getName();
+
+
+    static ZoneDraft ofCountries(final String name, final Set<CountryCode> countries, @Nullable final String description) {
+        final Set<Location> locations = countries.stream().map(country -> Location.of(country)).collect(toSet());
+        return new ZoneDraftImpl(name, locations, description);
     }
 
-    @Nullable
-    public String getDescription() {
-        return description;
+    static ZoneDraft of(final String name, final Set<Location> locations, @Nullable final String description) {
+        return new ZoneDraftImpl(name, locations, description);
     }
 
-    public Set<Location> getLocations() {
-        return locations;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public static ZoneDraft of(final String name, final Set<Location> locations) {
+    static ZoneDraft of(final String name, final Set<Location> locations) {
         return of(name, locations, null);
     }
 
-    public static ZoneDraft ofCountries(final String name, final Set<CountryCode> countries, @Nullable final String description) {
-        final Set<Location> locations = countries.stream().map(country -> Location.of(country)).collect(toSet());
-        return new ZoneDraft(name, locations, description);
-    }
-
-    public static ZoneDraft of(final String name, final Set<Location> locations, @Nullable final String description) {
-        return new ZoneDraft(name, locations, description);
-    }
 }
